http_proxy/requirements.txt
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
charset-normalizer==3.2.0
click==8.1.6
coverage==7.2.7
exceptiongroup==1.1.2
flake8==6.1.0
freezegun==1.2.2
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pycodestyle==2.11.0
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.0
pytest-aiohttp==0.3.0
pytest-cov==4.1.0
pytest-freezegun==0.4.2
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
yarl==1.9.2
